Ensure lockdir and rundir exist with correct permissions on startup
If /var/run or /var/lock happen to be on tmpfs filesystems, then the subdirectories
needed at runtime for condor-cron do not exist.  This commit adds code to the start()
function in /etc/rc.d/init.d/condor-cron to ensure the RUN and LOCK directories
are created and assigned the correct ownership.

Bug: https://ticket.grid.iu.edu/30906